Basic TCP connection done. Must to include x360ce driver

+++ x360ce.Engine/JocysCom/Drawing/Effects/Brightness.cs
@@ -1,4 +1,7 @@
+using System;
 using System.Drawing;
+using System.Drawing.Imaging;
+using System.Drawing.Drawing2D;
 
 namespace JocysCom.ClassLibrary.Drawing
 {
@@ -14,10 +17,10 @@ namespace JocysCom.ClassLibrary.Drawing
 			{
 				for (int x = 0; x < w; ++x)
 				{
-					byte[] p = ColorToBytes(b.GetPixel(x, y));
+					byte[] p = ColorToBytes(b.GetPixel(x, y)); 
 					for (int i = 1; i < 4; i++)
 					{
-						nVal = p[i] + level;
+						nVal = (int)(p[i] + level);
 						if (nVal < 0) nVal = 0;
 						if (nVal > 255) nVal = 255;
 						p[i] = (byte)nVal;
